Data Generator and Training code ready for execution
Data Generator:
The code in the Data generator has been optimized to load and computer faster.
For creating a batch of 32 for iterations of 32 reduced to 12 seconds for 1.2 minutes

Trainer:
The model has been updated to the correct dimensions as proposed in the paper.
The training loop is written, Waiting for Training.
BUGS: The code is not supporting Mixed precision at the moment because of the pytorch's
lack of support for mixed precision for LSTM's.
FEATURES: Will add mixed precision support once the release of pytorch 0.7.0 is finalized.

Future Plans:
Planning to update the Training loop from custom to PyTorch-Lightning.